Tidied up and redesigned:installing_apache_and_file.py

+++ logic/installing_apache_and_file.py
@@ -17,14 +17,15 @@ def copy_apache_and_exp(bit):
 def copy_file(source_dir, target_dir):
     if not os.path.exists(source_dir):
         return f"Исходная директория {source_dir} не существует."
-    else:
-        try:
-            shutil.copytree(source_dir, target_dir)
-            return f"Директория {source_dir} успешно скопирована в {target_dir}."
-        except FileExistsError:
-            return f"Целевая директория {target_dir} уже существует."
-        except Exception as e:
-            return f"Произошла ошибка: {e}"
+
+    if os.path.exists(target_dir):
+        return f"Целевая директория {target_dir} уже существует."
+
+    try:
+        shutil.copytree(source_dir, target_dir)
+        return f"Директория {source_dir} успешно скопирована в {target_dir}."
+    except Exception as e:
+        return f"Произошла ошибка: {e}"
 
 
 def move_and_rename_deskey_file(src_file_path, new_filename, destination_folder):
